Refactor DocLxCityCardWebServices.Codeunit.al to handle already redeemed city cards (#7276)
This pull request refactors the DocLxCityCardWebServices.Codeunit.al
file to handle already redeemed city cards. It adds a new case '523' to
the switch statement, which allows the code to acquire the previous
coupon and re-attempt redemption if it is already used. This ensures
that the redemption process is handled correctly and provides a better
user experience.