Auto/Self-Approve with comment (#7307)
Auto Approve will be triggered by posting a new comment on PR (not on PR
review).
Keywords: `auto-approve`, `auto approve`, `autoapprove`, `self-approve`,
`self approve` or `selfapprove`.
Comment needs to contain **only** the keyword (to avoid triggering
auto-approve if it's used in a sentence), case-insensitive, whitespaces
are trimmed.

By posting a new comment with one of keywords on a PR, workflow will be
triggered and the PR should be approved by the bot in about 30 seconds
(depending how many runners are available, sometimes it can take much
longer).