Get Global Vouchers before voucher search in SCAN_VOUCHER_2 action  (#7316)
SCAN_VOUCHER_2 actions tries to find the voucher type based on the
voucher reference. In order to do that it filters the vouchers on the
reference no and check if there is only one voucher with that reference
no. If there is only one voucher it uses it in the pos sale. If the pos
action finds more than one voucher with the same reference no or if it
fails to find any it opens a list with the available vouchers.
This was causing a problem when using the Global Vouchers, because you
will never have the voucher locally when you scan the refenrece no. The
current PR tries to retrieve the global voucher from the remote database
first and then proceeds with the the process flow.