POS Scenario Upgrade (#7312)
We replaced the FINISH_SALE and FINISH_CREDIT_SALE with a new engine.
When moving to the new engine we missed making some of the pos scenario
steps configurable. The current PR is trying to fix that.

New fields have been created on the "POS Audit Profile" to cater for
some of the pos scenario steps

**EjectPaymentBin** - new field has been created **"Bin Eject After
Sale"** on the **"POS Audit Profile"**
**PrintReceiptOnSale** - we depend on the **"Do Not Print Receipt on
Sale"** on the **"POS Audit Profile"**


New profile is created **"NPR MM POS Loyalty Profile"** that is
available on the POS Unit card. It is used for the configuration of the
Loyalty POS Scenarios.

**PointAssignmentOnSale** - is activated by the **"Assign Loyalty On
Sale"** field on **"NPR MM POS Loyalty Profile"**

New profile is created **"NPR TM POS Ticket Profile"** that is available
on the POS Unit card. It is used for the configuration of the Ticket POS
Scenarios.

**PrintTicketsOnSale** - is activated by the **"Print Ticket On Sale"**
field on the **"NPR TM POS Ticket Profile"**

New profile is created **"NPR MM POS Member Profile"** that is available
on the POS Unit card. It is used for the configuration of the the Member
POS Scenarios.
**PrintMembershipsOnSale** - is activated by the **"Print Membership On
Sale"** field on **"NPR MM POS Member Profile"**
**SendMemberNotificationOnSales** - is activated by the **"Send
Notification On Sale"** field on the **"NPR MM POS Member Profile"**

**DeliverCollectDocument** - new field has been created **"Deliver
Collect Document"** on **"NPR POS Sales Document Setup"** that enables
the functionality for all poses.


New procedures to move the POS Scenarios enable/disable setup to the
**"POS Audit Profile"**, **"NPR MM POS Member Profile"**, **"Deliver
Collect Document"**. Depending on the combinations the procedures create
new profiles with the necessary setups and assign the profiles to the
POSes.


Feature flag **posLifeCycleEventsWorkflowsEnabled_v2** has been removed
from the code.