Apply ServiceLocalTime for member arrival logic (#7322)
This pull request includes changes to apply the ServiceLocalTime for
member arrival logic. The logic has been updated to use the
ServiceLocalTime codeunit to get the local date and time for the member
arrival log entry. Additionally, the "Created Datetime" field in the
AdmittedTicketAccessEntry has been updated to use the current date and
time.