Add new events for customer creation process in membership create flow (#7323)
This pull request adds three new events to the membership create flow in
order to handle customer creation. The events are triggered before
assigning a customer number, after creating a customer, and after
creating a contact. These events provide hooks for custom logic and
allow for more flexibility in the membership creation process.