Update document no in archived retail voucher entries when posting a web sales order (#7339)
When a Web Sales Order that contains a Retail Voucher as a payment is
posted, the Retail Voucher Entry for the Voucher Payment gets the
Document No. of the Sales Order instead the one of the Posted Sales
Invoice. The current pr changes this behavior - now when the order is
posted the document no is going to be updated with the posted sales
invoice no.