Enforce validity period of exchange labels (#7342)
2 new fields on "Exchange Label Setup" page:
- "Enforce Exch. Validity Period": customers will need to enable this
feature, if they want to enforce exchange label validity period check;
- "Exchange Grace Period": a grace period (date formula). By specifying
the grace period, it will be possible to extend the period during which
the exchange label will still be accepted.
If the exchange label scan date exceeds the "Valid To" date + grace
period, an error message will be displayed and the exchange label won't
be accepted.


![image](https://github.com/navipartner/npcore/assets/50864934/c56d6b10-96d7-47a8-af81-e420b71438b2)