Create Feature Flag job in all companies (#7348)
Until now the Feature Flag job was created only in one company. Then we
were depending on a person to log into the environment which would start
the job. We had cases where the job was created in a not used company
and because of that nobody logged into it and the job was never started.
Because of that we have decided to to create the job in all companies
and let it run in each one. The current pr tries to achieve this.