Topic/sibl/hwd 367 customer display translation labels missing at paw (#7356)
Paw sko had issues with translations not beeing upadted on the User
Input page.

- Labels are now sent via InputRequest as well.
- Fixed The Prompt dialog on POS screen when receiving user input. When
Signature was not made, but phoneNo was input, it would not show phone
number.
- Added No Input label if neither Signature nor PhoneNo is present in
the input.
- When hitting the Login Screen we do not send a new "open display"
request (original issue for translations not beeing present if user had
clicked "Remake" to user input). Open Display and Download Media only
triggers when a new session is established, so if Minor Tom is exited,
or a user has clicked "Logout" in the settings in minor tom.