Decode enum and option value unicode escapes in replication (#7395)
In business central api 2.0 enum and option values are not a string
anymore, but strong typed values. The special characters and spaces in
those strongly typed values are encoded to unicode points hexadecimal
values.

For example in api 1.0 the customer price group option is looking like
this - Customer Price Group

In api 2.0 the customer price group option looks like this -
Customer_x0020_Price_x0020_Group

Code was added to search and decode unicode point values to text when
reading the api data.

In BC24 all custom apis are forced to use API2 by default. Before that
this was not the case.