Mix discount lots and combinations (#7409)
Since we're now getting only the mixed discount lines that are impacted
by the POS Sale the lot checkbox was not working correctly on the mixed
discounts. When enabled you have to have all items from the mixed
discount in the sale in order to active the discount. The changes in the
current PR check if all of the mixed discount lines impact the POS Sale
when the lot checkbox is set to true.