Add CaptureReservePoints functionality (#7423)
This pull request adds the functionality to capture reserved loyalty
points. It includes changes to the LoyaltyPointsMgrServer codeunit and
the XMLport "NPR MM CaptureReservePoints". The CaptureReservePoints
procedure is implemented to handle the capturing of reserved points and
update the import entry accordingly. This functionality allows for more
flexibility in managing loyalty points.