Update action code on pos button parameter values (#7433)
We have updated all of the pos menu buttons to use V3 actions. This
caused the follwoing problem - the action code on the shared parameters
between v2 and v3 actions was not updated with the V3 action code, but
stayed the same (V2 action code). This caused an issue when deleting the
old V2 action. The delete removed all of the parameters that were
related to it and caused the affected buttons to have missing
parameters. The current pr has an upgrade procedure that updates the
action codes on the pos button parameters.