Shopify and HeyLoyalty integration: more flexible task setup (#7452)
Hardcoded task processor codes for Shopify and HeyLoyalty integrations
did not allow the same codes to be used for alternative integrations in
PTE.
With this update we have removed code `SHOPIFY` and `HEYLOY`
reservations and made the functionality more flexible. We can now set
the codes individually in each company. Additionally, system won't
create task processors, if the integrations are not used in the company
(if the tables "NPR Spfy Integration Setup" and "NPR HL Integration
Setup" are empty).