Update captions and usage categories for member card and member list pages (#7470)
This pull request updates the captions and usage categories for the
member card and member list pages. The captions have been changed to
include "(Phone)" to indicate that these pages are specifically for
phone usage. Additionally, the usage categories have been changed to
"None" to remove these pages from the lists category. This improves the
clarity and organization of the pages.