"Refactor wf3 action and add member arrival response" (#7483)
Refactored the wf3 action to accept a JSON object instead of JSON text
in the workflow response. Added a response for member arrival that
contains member-related data to be shown in a toast message. Also
removed the regular "text message" and do not show the toast when the
backend shows a big confirm page.

Updated/added the MemberScanned toaster for all actions and added a
ToastMessageTimer duration parameter.

Duration of:
<0 means to do not show toaster
=0 means use default duration 15 seconds
>0 means show toaster for x amount of seconds