Issue multiple vouchers amount fix (#7501)
There was a problem with VAT when issuing vouchers that was resolved in
this PR - https://github.com/navipartner/npcore/pull/7435. The fix from
the PR didn't take into account that you can have one line on the POS
for a voucher with more than 1 quantity. The current PR addresses this
issue.