Suppress unnecessary bin transfer postings (#7525)
Bin transfer posting to G/L: the system now only posts amounts related
to the payment method for which the bin transfer was created.
Previously, the system would post float amounts for all payments
included in the checkpoint created for the bin transfer, regardless of
whether any a bin transfer amount was registered for the payment. This
confused users.