Search voucher by exact refrence no. (#7550)
Scan_Voucher_2 can find voucher types based on the reference number of
the voucher. When searching for the voucher type, we were filtering the
voucher table by the reference number and from there getting the voucher
type. We were using SETFILTER for this purpose. However, this caused a
problem for one of our customers because they used a very specific
pattern for generating the reference number that was beginning with ...
This was causing the SETFILTER to not find the specific reference number
but a list of vouchers. The current PR addresses this issue.