VIPPS Payment API URL switched to case insensitive (#7523)
Created based on:
https://linear.app/navipartner/issue/ISV2-365/vipps-payments-api-is-case-sensitive#comment-8b615e32

I do not have any special reason why it was initially created to always
convert Transaction ID to lowercase.
Everything was done to be in same format like other payment solutions
and VIPPS was already done for Munchmusset in their app, so I was doing
it based on those.

Now it is switched to be case insensitive, so issue reported in ISV2-365
will be resolved.