POS payment method items increase proposed amount fix (#7616)
When we have item groups assigned to a payment method we're calculating
the amount that can be paid with the payment method based on the setup.
After the calculation the salesperson can increase it manually. This was
not the case in the past. The issues was caused by the migration to
PaymentWF2. The current pr is trying to resolve the issue.