Public API codeunit for retrieving denominations (#7630)
A new public codeunit to access denominations from PTEs
The codeunit ID is 6184981 "NPR Public Denomination Access".
Use the `GetDenominations` procedure. You'll need to specify a value for
the `AttachedToRecordID` (the record ID of the record for which you want
to get denominations). If the record is "NPR POS Payment Bin Checkp.",
you'll also need to provide a value for the `AttachedToID` parameter.
The procedure will return the denominations as a set of records using
the temporary table 6150910 "NPR Entry Denomination" as a container.