Inventory cost adjmt job: allow changing parameter string (#7636)
Some customers don't want the reports generated by the job to be saved
in their inbox.
The update will allow changing the parameter string for the job and the
system will no longer enforce the default parameter string for the job
when it refreshes the job list on full user logins.