Endless Aisle Manual Capture Sales Documents (#7568)
This PR contains the changes for the Endless Aisle Adyen Integration. 

"POS EFT Payment Reservation Setup" new page where you can set up which
gateway is going to be used for the payment reservation and the G/L
Account to which the payment is going to be posted.

STORE Shipment Profile - a profile where you can specify the different
shipping fees from which you can choose when making the payment
reservation.

"Manual Capture" - new field on the EFT Setup Paraemeters - when checked
the payment on the POS will only authorise the funds, but they won't be
captured.

EFT_RESERVE_DOC_PAY - new action that can be used to make payment
reservations on the POS for already existing sales documents.

SALES_DOC_EXP - new parameters to the pos action
 - POSPaymentReservation - enables the payment reservation 
- POSPaymentMethodCode - the payment method code which is going to be
used to make the payment reservation.
 - SelectShipmentMethod - enables the shipment fee selection

---------

Co-authored-by: Mihails Pojasnikovs <mipo@navipartner.com>
Co-authored-by: Nikolina Duvnjak <nidu@navipartner.com>
Co-authored-by: Nikolina Duvnjak <147609865+nidu1701@users.noreply.github.com>