POS action RV_NEW_WAITER_PAD: possibility to hide unnecessary parameters (#7639)
It was not possible to set whether the system should ask for customer
information (name, phone, email) as well as the number of guests, when
creating a new waiter pad for a table in the restaurant view. The system
would always show a popup with all the request fields.
Fixed in this PR.