Double rounding bug fix (#7679)
We recently had an issue where customer got 2 same rounding lines after
ending the sale, thus could not post the entry because of the amount
difference of the 2nd rounding line
This PR is the fix for it. I'm deleting the existing rounding lines
before the only one that should be on there, is calculated.