Shopify integration: removed click and collect feature (#7680)
The functionality was originally developed for Vibholm, and I don't
think we should have included it in Np Core in the first place. The
functionality has several drawbacks:
- It requires a custom app to be installed on Shopify side. The app has
been developed by a third party and we don't have access to it.
- The functionality is very limited: it doesn't leave any traces on the
Shopify side (no Shopify orders are created), it only supports payment
upon receipt of the goods in a store.

Now that we have decided to involve the Magento and Frontend teams to
extend our Shopify integration, we can create an alternative and much
richer Click&Collect process with Shopify.