Best price calculation adjustments (#7685)
A customer informed us that in the specific case described in the
connected case it was not possible to select the customer. The problem
was a combination of several issue, but the main one from where
everything started was the best price calculation. It's a complex case
that is difficult to explain but it boils down to the best price
calculation which takes the discount into account when it runs. So if
you have two different prices that always get 100% the best price
calculation will always take the first one. This was causing the best
price calculation to calculate different prices when the discount was
calculated which was causing the POS line to split into two lines,
because the unit price before the discount calculation was different
than the unit price after the calculation. What should have happened is
the price calculation should have always calculated the same unit price
and shouldn't have split the pos line into different lines.