Added new features to feature management (#7634)
Added 2 new features:
Scenarios obsoleted - it needs to be set to true on all new tenants. It
will hide all pages that are related to the POS Scenarios Functionality.
New POS Editor - it needs to be set to true on all new tenants. It will
hide all pages that are related to the old POS Editor.

***Note for reviewer***: I am open to further suggestions in terms of
changing some naming of application area or captions, as well as new
pages that we should obsolete