Added users email to callback (#7622)
Added authenication email and contact email addresses to callback in
order to have option in case system to identify NaviPartner users in BC
SaaS Users list.

Initially we are going to use Authentication email only.
So all accounts with @navipartner.com or @navipartner.dk or Navi Partner
Technician will set user as Delegated user and that one will not be
included in user logins counting.

Or differently explained, we will now that users like these ones: 


![image](https://github.com/user-attachments/assets/8f8a363b-ab37-41f6-916a-93b7690512c3)

are actually NaviPartner users. Or even if they are not, we will know
who they are, based on email address.